refactor(user-domain): user 엔티티 구조 정리 및 status 필드 수정
- user 관련 예외 클래스를 com.pado.user.exception 패키지로 이동
- status 필드 삭제
